producer/requirements.txt
websocket-client
kafka-python